Made new folder code and moved games in code folder.
